STOR-31 moved Services
